Strict null checks -- begin untangling dependency cycle (#8338)
* getToken/refreshToken

* src/background/requests.ts

* src/contentScript/pageEditor/elementPicker.ts

* src/background/partnerIntegrations.ts

* whoops

* cleanup

* fix widget and getScopeAndId

* auto-add

* src/extensionConsole/pages/mods/GetStartedView.tsx

* src/mods/useModViewItems.tsx

* test memo compare dependencies

* fix scopeAndId tests

* autosuggest

* src/background/backgroundPlatform.ts

* src/background/restrictUnauthenticatedUrlAccess.ts

* src/bricks/effects/insertHtml.ts

* ./bricks/effects/runSubTour.ts

* ./bricks/effects/scrollIntoView.ts

* src/bricks/transformers/controlFlow/WithAsyncModVariable.ts

* mod state / variable context

* src/bricks/transformers/RunMetadataTransformer.ts

* src/bricks/transformers/splitText.ts

* src/components/fields/schemaFields/widgets/varPopup/useTreeRow.ts

* src/components/walkthroughModal/WalkthroughModalApp.tsx

* src/extensionConsole/pages/brickEditor/BrickHistory.tsx

* fix test

* src/extensionConsole/pages/mods/utils/exportBlueprint.ts

* src/extensionConsole/pages/mods/utils/exportBlueprint.ts

* src/pageScript/elementInfo.ts

* src/sidebar/LoginPanel.tsx

* src/testUtils/factories/authFactories.ts

* src/bricks/effects/attachAutocomplete.ts

* more bricks

* fix test

* src/components/fields/schemaFields/fieldTestUtils.ts

* add a bunch more

* the rest

* cleanup

* fix lint

* wip

* remove file

* remove logs

* fix factory

* update counts

* extensionPagePlatform

* partnerIntegrations

* update counts

* finding cycles

* documenting cycle

* move starterbricks into separate folders

* extract shared types from contextmenu

* migrate shared types for menuItemExtension

* migrate panel types; migrate more contextMenu types

* migrate more menuitem types

* migrate quickbar types

* migrate quickbar provider types

* migrate sidebar types

* migrate tour types

* update file paths

* auto-add

* fix types

* auto-add

* remove outdated cycle comments

---------
