+++ src/contentScript/messenger/registration.ts
@@ -30,31 +30,31 @@ import {
   queueReactivateTab,
   reactivateTab,
   removePersistedExtension,
-} from "@/contentScript/lifecycle"; // 219 strictNullCheck errors
-import { insertPanel } from "@/contentScript/pageEditor/insertPanel"; // 255 strictNullCheck errors
-import { insertButton } from "@/contentScript/pageEditor/insertButton"; // 261 strictNullCheck errors
+} from "@/contentScript/lifecycle"; // 214 strictNullCheck errors
+import { insertPanel } from "@/contentScript/pageEditor/insertPanel"; // 216 strictNullCheck errors
+import { insertButton } from "@/contentScript/pageEditor/insertButton"; // 222 strictNullCheck errors
 import {
   clearDynamicElements,
   disableOverlay,
   enableOverlay,
   runExtensionPointReader,
   updateDynamicElement,
-} from "@/contentScript/pageEditor/dynamic"; // 228 strictNullCheck errors
+} from "@/contentScript/pageEditor/dynamic"; // 217 strictNullCheck errors
 import {
   runBlockPreview,
   resetTab,
   runRendererBlock,
-} from "@/contentScript/pageEditor"; // 258 strictNullCheck errors
-import { runBrick } from "@/contentScript/executor"; // 134 strictNullCheck errors
-import selectElement from "@/contentScript/pageEditor/selectElement"; // 235 strictNullCheck errors
+} from "@/contentScript/pageEditor"; // 219 strictNullCheck errors
+import { runBrick } from "@/contentScript/executor"; // Depends on background/messenger to pass strictNullCheck
+import selectElement from "@/contentScript/pageEditor/selectElement"; // 196 strictNullCheck errors
 import {
   runHeadlessPipeline,
   runMapArgs,
   runRendererPipeline,
 } from "@/contentScript/pipelineProtocol"; // Depends on background/messenger to pass strictNullCheck
-import { reloadActivationEnhancements } from "@/contentScript/loadActivationEnhancementsCore"; // 194 strictNullCheck errors
-import { getAttributeExamples } from "@/contentScript/pageEditor/elementInformation"; // 192 strictNullCheck errors
-import { getCopilotHostData } from "@/contrib/automationanywhere/SetCopilotDataEffect"; // 19 strictNullCheck errors
+import { reloadActivationEnhancements } from "@/contentScript/loadActivationEnhancementsCore"; // 187 strictNullCheck errors
+import { getAttributeExamples } from "@/contentScript/pageEditor/elementInformation"; // 185 strictNullCheck errors
+import { getCopilotHostData } from "@/contrib/automationanywhere/SetCopilotDataEffect"; // Depends on background/messenger to pass strictNullCheck
 import { showBannerFromConfig } from "@/contentScript/integrations/deferredLoginController"; // Depends on background/messenger to pass strictNullCheck
 
 expectContext("contentScript");
