+++ src/pageEditor/starterBricks/menuItem.ts
@@ -31,11 +31,7 @@ import {
   selectIsAvailable,
 } from "@/pageEditor/starterBricks/base";
 import { omitEditorMetadata } from "./pipelineMapping";
-import {
-  type MenuDefinition,
-  type MenuItemStarterBrickConfig,
-  MenuItemStarterBrickABC,
-} from "@/starterBricks/menuItemExtension";
+import { MenuItemStarterBrickABC } from "@/starterBricks/menuItem/menuItemExtension";
 import { type StarterBrickConfig } from "@/starterBricks/types";
 import { identity, pickBy } from "lodash";
 import { getDomain } from "@/permissions/patterns";
@@ -48,6 +44,10 @@ import {
   type ButtonSelectionResult,
 } from "@/contentScript/pageEditor/types";
 import { type ActionFormState } from "./formStateTypes";
+import {
+  type MenuItemDefinition,
+  type MenuItemStarterBrickConfig,
+} from "@/starterBricks/menuItem/types";
 
 function fromNativeElement(
   url: string,
@@ -87,7 +87,7 @@ function fromNativeElement(
 
 function selectExtensionPointConfig(
   formState: ActionFormState,
-): StarterBrickConfig<MenuDefinition> {
+): StarterBrickConfig<MenuItemDefinition> {
   const { extensionPoint } = formState;
   const {
     definition: {
@@ -140,7 +140,7 @@ async function fromExtension(
   config: ModComponentBase<MenuItemStarterBrickConfig>,
 ): Promise<ActionFormState> {
   const extensionPoint = await lookupExtensionPoint<
-    MenuDefinition,
+    MenuItemDefinition,
     MenuItemStarterBrickConfig,
     "menuItem"
   >(config, "menuItem");
