+++ src/starterBricks/menuItem/menuItemExtension.ts
@@ -33,17 +33,14 @@ import {
 import {
   StarterBrickABC,
   type StarterBrickConfig,
-  type StarterBrickDefinition,
 } from "@/starterBricks/types";
 import { type Metadata } from "@/types/registryTypes";
 import { type Permissions } from "webextension-polyfill";
 import reportEvent from "@/telemetry/reportEvent";
 import { Events } from "@/telemetry/events";
-import { type MessageConfig } from "@/utils/notify";
 import { getNavigationId } from "@/contentScript/context";
 import getSvgIcon from "@/icons/getSvgIcon";
 import { selectEventData } from "@/telemetry/deployments";
-import { type BrickConfig, type BrickPipeline } from "@/bricks/types";
 import apiVersionOptions, {
   DEFAULT_IMPLICIT_TEMPLATE_ENGINE,
 } from "@/runtime/apiVersionOptions";
@@ -63,7 +60,6 @@ import {
 } from "@/errors/businessErrors";
 import { PromiseCancelled } from "@/errors/genericErrors";
 import { rejectOnCancelled } from "@/errors/rejectOnCancelled";
-import { type IconConfig } from "@/types/iconTypes";
 import { type Schema } from "@/types/schemaTypes";
 import { type ResolvedModComponent } from "@/types/modComponentTypes";
 import { type Brick } from "@/types/brickTypes";
@@ -87,73 +83,17 @@ import {
 import type { PlatformProtocol } from "@/platform/platformProtocol";
 import { DEFAULT_ACTION_RESULTS } from "@/starterBricks/starterBrickConstants";
 import { propertiesToSchema } from "@/utils/schemaUtils";
-
-interface ShadowDOM {
-  mode?: "open" | "closed";
-  tag?: string;
-}
-
-/**
- * @since 1.7.8
- */
-export type AttachMode =
-  // Add menu items once. If a menu item is removed, PixieBrix will still attempt to re-add it.
-  | "once"
-  // Watch for new menus on the screen and add menu items to them
-  | "watch";
+import {
+  type MenuItemStarterBrickConfig,
+  type AttachMode,
+  type MenuItemDefinition,
+  type MenuTargetMode,
+} from "@/starterBricks/menuItem/types";
 
 const DATA_ATTR = "data-pb-uuid";
 
 const MENU_INSTALL_ERROR_DEBOUNCE_MS = 1000;
 
-export type MenuItemStarterBrickConfig = {
-  /**
-   * The button caption to supply to the `caption` in the extension point template.
-   * If `dynamicCaption` is true, can include template expressions.
-   */
-  caption: string;
-
-  /**
-   * (Optional) the icon to supply to the icon in the extension point template
-   */
-  icon?: IconConfig;
-
-  /**
-   * The action to perform when the button is clicked
-   */
-  action: BrickConfig | BrickPipeline;
-
-  /**
-   * (Experimental) condition to determine whether to show the menu item
-   * @see if
-   */
-  if?: BrickConfig | BrickPipeline;
-
-  /**
-   * True if caption is determined dynamically (using the reader and templating)
-   */
-  dynamicCaption?: boolean;
-
-  /**
-   * True to prevent button to be clicked again while action is in progress
-   */
-  synchronous: boolean;
-
-  /**
-   * (Experimental) message to show on error running the extension
-   */
-  onError?: MessageConfig;
-  /**
-   * (Experimental) message to show if the user cancelled the action (e.g., cancelled a form, or the Cancel brick ran)
-   */
-  onCancel?: MessageConfig;
-  /**
-   * (Experimental) message to show on success when running the extension
-   */
-  onSuccess?: MessageConfig | boolean;
-};
-
-// eslint-disable-next-line local-rules/persistBackgroundData -- Static
 const actionSchema: Schema = {
   oneOf: [
     { $ref: "https://app.pixiebrix.com/schemas/effect#" },
@@ -171,8 +111,6 @@ async function cancelOnNavigation<T>(promise: Promise<T>): Promise<T> {
   const isNavigationCancelled = () => getNavigationId() !== startNavigationId;
   return rejectOnCancelled(promise, isNavigationCancelled);
 }
-
-type MenuTargetMode = "document" | "eventTarget";
 
 export abstract class MenuItemStarterBrickABC extends StarterBrickABC<MenuItemStarterBrickConfig> {
   /**
@@ -817,71 +755,12 @@ export abstract class MenuItemStarterBrickABC extends StarterBrickABC<MenuItemSt
   }
 }
 
-interface MenuDefaultOptions {
-  caption?: string;
-  [key: string]: string;
-}
-
-export type MenuPosition =
-  | "append"
-  | "prepend"
-  | {
-      // Element to insert the menu item before, selector is relative to the container
-      sibling: string | null;
-    };
-
-/**
- * @since 1.7.16
- */
-
-export interface MenuDefinition extends StarterBrickDefinition {
-  type: "menuItem";
-  /**
-   * The HTML template to render the button/menu item.
-   */
-  template: string;
-  /**
-   * Position in the menu to insert the item.
-   */
-  position?: MenuPosition;
-  /**
-   * Selector targeting the menu location
-   */
-  containerSelector: string;
-  /**
-   * Selector passed to `.parents()` to determine the reader context. Must match exactly one element.
-   * See https://api.jquery.com/parents/
-   * @deprecated use targetMode and the Traverse Elements brick instead
-   */
-  readerSelector?: string;
-  /**
-   * The element to pass as the root to the readers and extension (default="document")
-   * @since 1.7.16
-   * @see readerSelector
-   */
-  targetMode?: MenuTargetMode;
-  /**
-   * Wrap menu item in a shadow DOM
-   * @deprecated do we still want to support this? Is it used anywhere?
-   */
-  shadowDOM?: ShadowDOM;
-  /**
-   * Default options for ModComponentBases attached to the extension point
-   */
-  defaultOptions?: MenuDefaultOptions;
-  /**
-   * Mode for attaching the menu to the page. Defaults to "once"
-   * @since 1.7.28
-   */
-  attachMode?: AttachMode;
-}
-
 export class RemoteMenuItemExtensionPoint extends MenuItemStarterBrickABC {
-  private readonly _definition: MenuDefinition;
+  private readonly _definition: MenuItemDefinition;
 
   public readonly permissions: Permissions.Permissions;
 
-  public readonly rawConfig: StarterBrickConfig<MenuDefinition>;
+  public readonly rawConfig: StarterBrickConfig<MenuItemDefinition>;
 
   public override get defaultOptions(): {
     caption: string;
@@ -896,7 +775,7 @@ export class RemoteMenuItemExtensionPoint extends MenuItemStarterBrickABC {
 
   constructor(
     platform: PlatformProtocol,
-    config: StarterBrickConfig<MenuDefinition>,
+    config: StarterBrickConfig<MenuItemDefinition>,
   ) {
     // `cloneDeep` to ensure we have an isolated copy (since proxies could get revoked)
     const cloned = cloneDeep(config);
@@ -1055,7 +934,7 @@ export class RemoteMenuItemExtensionPoint extends MenuItemStarterBrickABC {
 
 export function fromJS(
   platform: PlatformProtocol,
-  config: StarterBrickConfig<MenuDefinition>,
+  config: StarterBrickConfig<MenuItemDefinition>,
 ): StarterBrick {
   const { type } = config.definition;
   if (type !== "menuItem") {
