+++ src/store/checkAvailableInstalledExtensions.test.ts
@@ -23,21 +23,17 @@ import { type ModComponentsRootState } from "@/store/extensionsTypes";
 import { selectModComponentAvailability } from "@/pageEditor/slices/editorSelectors";
 import { getInstalledExtensionPoints } from "@/contentScript/messenger/api";
 import { validateRegistryId } from "@/types/helpers";
-import {
-  type MenuDefinition,
-  RemoteMenuItemExtensionPoint,
-} from "@/starterBricks/menuItemExtension";
+import { RemoteMenuItemExtensionPoint } from "@/starterBricks/menuItem/menuItemExtension";
 import { type StarterBrickConfig } from "@/starterBricks/types";
 import { type Metadata } from "@/types/registryTypes";
-import {
-  type QuickBarDefinition,
-  RemoteQuickBarExtensionPoint,
-} from "@/starterBricks/quickBarExtension";
+import { RemoteQuickBarExtensionPoint } from "@/starterBricks/quickBar/quickBarExtension";
 import { starterBrickConfigFactory } from "@/testUtils/factories/modDefinitionFactories";
 import { standaloneModDefinitionFactory } from "@/testUtils/factories/modComponentFactories";
 import { metadataFactory } from "@/testUtils/factories/metadataFactory";
 import { getCurrentInspectedURL } from "@/pageEditor/context/connection";
 import { getPlatform } from "@/platform/platformContext";
+import { type MenuItemDefinition } from "@/starterBricks/menuItem/types";
+import { type QuickBarDefinition } from "@/starterBricks/quickBar/types";
 
 jest.mock("@/contentScript/messenger/api");
 
@@ -73,7 +69,7 @@ describe("checkAvailableInstalledExtensions", () => {
           id: availableButtonId,
         });
       },
-      definition(): MenuDefinition {
+      definition(): MenuItemDefinition {
         return {
           type: "menuItem",
           containerSelector: "",
@@ -84,7 +80,7 @@ describe("checkAvailableInstalledExtensions", () => {
           reader: validateRegistryId("@pixiebrix/document-context"),
         };
       },
-    }) as StarterBrickConfig<MenuDefinition>;
+    }) as StarterBrickConfig<MenuItemDefinition>;
     const availableButtonExtensionPoint = new RemoteMenuItemExtensionPoint(
       getPlatform(),
       availableButtonStarterBrickConfig,
